requirements.txt
pillow
colorthief
cgi
